Update makefile, still using g++-9 for initializer_list. Fix -Wall warnings.

+++ src/Comparison.hpp
@@ -13,7 +13,7 @@ public:
     void compare_suffix_tree();
     SuffixTree build_suffix_tree(std::string genome);
     std::list<std::string> remove_repeating_nodes(std::list<std::string>);
-    std::list<std::string> remove_small_nodes(std::list<std::string>, int size);
+    std::list<std::string> remove_small_nodes(std::list<std::string>, size_t size);
     std::list<SuffixNode> remove_substrings(std::list<SuffixNode>);
     int check_if_substring(std::string str, std::list<SuffixNode>);
